@colesbury 2019-12-03
1. I think this code path is used by GPU EmbeddingBag backward as well. Can you benchmark that to make sure there's no regression?

2. I'm a little concerned about the regression for the 64x64 contiguous case. Can you investigate this a bit more? For example, keep the columns of `input` fixed as 64 but increase the number of rows. (64 is a plausible number of columns but seems really small for the number of rows)