@bddppq 2019-12-04
@colesbury Good point. I re-ran the benchmarks for more different combinations of (#row, #col) and found for contiguous input with small #col, `index_select` is consistently faster (the tipping point is 64), for other cases `index` is faster.

Benchmark code: https://gist.github.com/bddppq/216fc8abfa8b68fa6b861974e977ded6

![image](https://user-images.githubusercontent.com/9300575/70105141-e405a480-15f4-11ea-95e2-8c26fa4188e1.png)
![image](https://user-images.githubusercontent.com/9300575/70105150-e962ef00-15f4-11ea-8fee-26bf155228a2.png)
![image](https://user-images.githubusercontent.com/9300575/70105161-f253c080-15f4-11ea-8d05-227c4b6cc847.png)
![image](https://user-images.githubusercontent.com/9300575/70105180-fe3f8280-15f4-11ea-8d07-b4741a108638.png)
![image](https://user-images.githubusercontent.com/9300575/70105231-131c1600-15f5-11ea-9fa8-5d5825167f77.png)
![image](https://user-images.githubusercontent.com/9300575/70105206-08618100-15f5-11ea-9dcf-6b18e67c3c40.png)
![image](https://user-images.githubusercontent.com/9300575/70105245-1adbba80-15f5-11ea-8ac7-38328ecd1ad8.png)
![image](https://user-images.githubusercontent.com/9300575/70105283-3cd53d00-15f5-11ea-90f7-1ea0dbdffebf.png)
![image](https://user-images.githubusercontent.com/9300575/70105292-41015a80-15f5-11ea-94f1-a6a84e2d97ea.png)
![image](https://user-images.githubusercontent.com/9300575/70105297-452d7800-15f5-11ea-889a-9215a0618bde.png)
![image](https://user-images.githubusercontent.com/9300575/70105302-48286880-15f5-11ea-9ced-e4913c638ee3.png)
![image](https://user-images.githubusercontent.com/9300575/70105315-4bbbef80-15f5-11ea-8c75-c6d0e7940c67.png)
![image](https://user-images.githubusercontent.com/9300575/70105318-4fe80d00-15f5-11ea-9afc-45e2b9c540fa.png)
![image](https://user-images.githubusercontent.com/9300575/70105355-6aba8180-15f5-11ea-9d9d-830df5fe93dd.png)
![image](https://user-images.githubusercontent.com/9300575/70105360-6e4e0880-15f5-11ea-8431-886e64734327.png)
![image](https://user-images.githubusercontent.com/9300575/70105364-7148f900-15f5-11ea-9b40-9810b44cf43b.png)
![image](https://user-images.githubusercontent.com/9300575/70105368-74dc8000-15f5-11ea-9887-4c8e142322ea.png)
![image](https://user-images.githubusercontent.com/9300575/70105376-79a13400-15f5-11ea-87e5-b0e311dc55dd.png)